+++ DemoApplication/VisitorApplication/src/SwitchCyclicAcyclicVisitables.h
@@ -1,0 +1,15 @@
+/*
+ * SwitchCyclicAcyclicVisitables.h
+ *
+ *  Created on: 05.12.2016
+ *      Author: Gerd
+ */
+
+#ifndef SWITCHCYCLICACYCLICVISITABLES_H_
+#define SWITCHCYCLICACYCLICVISITABLES_H_
+
+
+
+
+
+#endif /* SWITCHCYCLICACYCLICVISITABLES_H_ */
