+++ DemoApplication/VisitorApplication/src/SwitchCyclicAcyclicVisitors.h
@@ -1,0 +1,15 @@
+/*
+ * SwitchCyclicAcyclicVisitors.h
+ *
+ *  Created on: 05.12.2016
+ *      Author: Gerd
+ */
+
+#ifndef SWITCHCYCLICACYCLICVISITORS_H_
+#define SWITCHCYCLICACYCLICVISITORS_H_
+
+
+
+
+
+#endif /* SWITCHCYCLICACYCLICVISITORS_H_ */
